1.Fix: player dead bug and areatrigger bug. 2.cleanup the use string format, always use "%u" if the value is unsigned. 3.all the save and load of player guid now use uint32 type. because use "%d" and "%lu" or "%ld" in function PExecute and PQuery sometime get error format string. 4.change command cdist as distance to show the distance of player to creature. 5.change the calculate of GetDistanceSq. 6.add config data in World.cpp to save config data.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@1192 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/ObjectMgr.h
@@ -211,7 +211,7 @@ class ObjectMgr
 
         PlayerCreateInfo* GetPlayerCreateInfo(uint32 race, uint32 class_);
 
-        uint64 GetPlayerGUIDByName(const char *name) const;
+        uint32 GetPlayerGUIDByName(const char *name) const;
         bool GetPlayerNameByGUID(const uint64 &guid, std::string &name) const;
 
         bool GetGlobalTaxiNodeMask( uint32 curloc, uint32 *Mask );
